refactor: move serializers to dedicated module
